requirements.txt
streamlit==1.32.2
google-generativeai==0.5.0
python-dotenv==1.0.1
